# libexpr: Actually cache line information in PosTable